+++ Documentation/SBV/Examples/ADT/Expr.hs
@@ -94,8 +94,9 @@ eval = go []
        get :: SList (String, Integer) -> SString -> SInteger
        get = smtFunction "get"
            $ \env s -> [sCase| env of
-                          []          -> 0
-                          (k, v) : es -> ite (s .== k) v (get es s)
+                          []                    -> 0
+                          (k, v) : es | s .== k -> v
+                                      | True    -> get es s
                        |]
 
 -- | A basic theorem about 'eval'.